requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
openai>=1.0.0
nltk>=3.8.0
websockets>=11.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiosqlite>=0.19.0
pyyaml>=6.0.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
